move json-rpc protocol to its own package
