+++ neptulon/neptulon.go
@@ -17,9 +17,12 @@ type App struct {
 	connMutex  sync.Mutex
 }
 
-// NewApp creates and returns a new Neptulon app. This is the default TLS constructor.
+// NewApp creates a new Neptulon app. This is the default TLS constructor.
 // Debug mode dumps raw TCP data to stderr (log.Println() default).
 func NewApp(cert, privKey []byte, laddr string, debug bool) (*App, error) {
+
+	// todo: App.UseTLS(cert, priv) should be optional and not in the constructor
+
 	l, err := Listen(cert, privKey, laddr, debug)
 	if err != nil {
 		return nil, err
